requirements.txt

flask
pytesseract
pillow
pandas
